+++ frontend/src/components/TabTurno.jsx
@@ -11,6 +11,9 @@ import {useTurnos, deleteTurnos, addTurnos, modifyTurnos, getTurnosQuery} from "
 import { usePolitica } from '../hooks/usePolitica.js';
 import { useCentros } from '../hooks/useCentros.js';
 import { useTiposAnalisis } from '../hooks/useTiposAnalisis.js';
+import TabTurnoGestion from './tabsUsers/TabTurnoGestion';
+import TabTurnoRegistrar from './tabsUsers/TabTurnoRegistrar';
+import TabTurnoResultados from './tabsUsers/TabTurnoResultados';
 
 const generateTimeSlots = (horaInicio, horaFin, intervalo) => {
   const horarios = [];
@@ -69,13 +72,13 @@ const { register: registerAdd, handleSubmit: handleSubmitAdd, formState: { error
 
   const handleResultadosClick = (id) => {
     setResultadosId(id);
-    console.log("Resultados del turno seleccionado:", turnosPaciente.find((t) => t.id === Number(id)).resultados);
     setShowModal(true);
   };
 
   const handleCerrarModal = () => {
     setShowModal(false);
     setTurnoAEliminarId(null);
+    setResultadosId(null);
   };
 
   const handleConfirmarEliminacion = async () => {
@@ -141,7 +144,6 @@ useEffect(() => {
     );
 
     setTurnosFiltradosResultados(turnosResultados);
-    console.log("Turnos filtrados para resultados:", turnosFiltradosResultados);
   }
   if (!Array.isArray(turnosPaciente) || turnosPaciente.length === 0) {
     setTurnosFiltradosGestion([]);
@@ -161,334 +163,13 @@ useEffect(() => {
       justify
     >
       <Tab eventKey="gestiondeturnos" title="Gestión de Turnos">
-        <h2 className='titulo'>Mis turnos</h2>
-        {turnosFiltradosGestion.length > 0 ? (
-          <div style={{ marginTop: '20px', overflowX: 'scroll' }}>
-            <table className="table">
-              <thead>
-                <tr>
-                  <th>Número de Turno</th>
-                  <th>Tipo de Análisis</th>
-                  <th>Centro de Atención</th>
-                  <th>Fecha y Hora</th>
-                  <th>Estado</th>
-                  <th>Observación</th>
-                  <th>Recibe Mail</th>
-                  <th>Cancelar Turno</th>
-                </tr>
-              </thead>
-              <tbody style={{ verticalAlign: 'middle' }}>
-                {turnosFiltradosGestion.map((turno) => {
-                  if (turno.estado === "Anulado") {
-                    return null;
-                  }
-                  return (
-                    <tr key={turno.id}>
-                      <td>{turno.id}</td>
-                      <td>{turno.tipoAnalisis?.nombre || '-'}</td>
-                      <td>{turno.centroAtencion?.nombre || '-'}</td>
-                      <td>{new Date(turno.fechaHoraReserva).toLocaleString()}</td>
-                      <td>{turno.estado}</td>
-                      <td>{turno.observacion === "" ? "-" : turno.observacion}</td>
-                      <td>{turno.recibeMail ? "Sí" : "No"}</td>
-                      <td>
-                        <button
-                          onClick={() => handleEliminarClick(turno.id)}
-                          style={{
-                            background: 'none',
-                            border: 'none',
-                            color: 'red',
-                            cursor: 'pointer',
-                            fontSize: '20px',
-                            fontStyle: 'underline'
-                          }}
-                        >
-                          ❌
-                        </button>
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          </div>
-        ) : (
-          <p style={{ marginTop: "30px" }}>No tienes turnos registrados.</p>
-        )}
-        {showModal && (
-          <div style={{
-            position: 'fixed',
-            top: 0,
-            left: 0,
-            width: '100%',
-            height: '100%',
-            backgroundColor: 'rgba(0, 0, 0, 0.5)',
-            display: 'flex',
-            justifyContent: 'center',
-            alignItems: 'center',
-            zIndex: 1000,
-          }}>
-            <div style={{
-              backgroundColor: 'white',
-              padding: '20px',
-              borderRadius: '5px',
-              textAlign: 'center',
-              minWidth: '300px'
-            }}>
-              <h4 style={{fontWeight: 'bold', fontFamily: 'Segoe UI, Tahoma, Geneva, Verdana, sans-serif'}}>
-                Confirmar Cancelación</h4>
-              <p>
-                ¿Estás seguro de que quieres cancelar este turno?</p>
-              <div style={{ marginTop: '20px' }}>
-                <button onClick={handleCerrarModal} className='login-btn' style={{ backgroundColor: 'red' }}>
-                  Volver
-                </button>
-                 <button onClick={handleConfirmarEliminacion} className='login-btn' style={{ marginLeft: '10px' }}>
-                  Confirmar
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
+        <TabTurnoGestion turnosFiltradosGestion={turnosFiltradosGestion} onEliminarClick={handleEliminarClick} />
       </Tab>
       <Tab eventKey="registrarturno" title="Registrar Turno">
-        <h2 className='titulo'>Registrar Turno</h2>
-        <form
-          encType="multipart/form-data"
-          className="login-formReg"
-          onSubmit={handleSubmitAdd(onSubmitAdd)}
-          noValidate
-        >
-          <div className="form-group" id="uno">
-            <label htmlFor="text">Tipo de Análisis</label>
-            <select
-              id="tipoAnalisis"
-              {...registerAdd("tipoAnalisis", {
-                required: "Tipo de Análisis requerido",
-              })}
-              className="form-input"
-            >
-              <option value="">-</option>
-              {tipos.map((ta, index) => (
-                <option key={index} value={ta.id}>
-                  {ta.id} - {ta.nombre}
-                </option>
-              ))}
-            </select>
-            {errorsAdd.tipoAnalisis && (
-              <div className="error-message">{errorsAdd.tipoAnalisis.message}</div>
-            )}
-          </div>
-
-          <div id="dos" className="form-group">
-            <label htmlFor="date">Fecha del Turno</label>
-            <input
-              type="date"
-              id="fechaHoraReserva"
-              {...registerAdd("fechaHoraReserva", {
-                required: "Fecha requerida",
-                validate: (value) => {
-                  const selectedDate = new Date(value);
-                  const currentDate = new Date();
-                  currentDate.setHours(0, 0, 0, 0);
-                  if (selectedDate < currentDate) {
-                    return "Fecha de turno inválida";
-                  }
-                  return true;
-                },
-              })}
-              className="form-input"
-            />
-            {errorsAdd.fechaHoraReserva && (
-              <div className="error-message">{errorsAdd.fechaHoraReserva.message}</div>
-            )}
-            </div>
-              <div id="tres" className="form-group">
-            {(fechaHoraReserva && errorsAdd.fechaHoraReserva == undefined) && (
-              <>
-                <label htmlFor="time">Hora del Turno</label>
-                <select 
-                  id="horaReserva"
-                  {...registerAdd("horaReserva", {
-                    required: "Hora requerida",
-                  })}
-                  className="form-input"
-                >
-                  <option value="">-</option>
-                  {horariosDisponibles.map((hora, index) => (
-                    <option key={index} value={hora}>
-                      {hora}
-                    </option>
-                  ))}
-                </select>
-                {errorsAdd.horaReserva && (
-                  <div className="error-message">{errorsAdd.horaReserva.message}</div>
-                )}
-              </>
-            )}
-          </div>
-
-          <div className="form-group" style={{gridColumn: "1", gridRow : "2"}} id="tres">
-            <label htmlFor="text">Centro de Atención</label>
-            <select
-              id="centroAtencion"
-              {...registerAdd("centroAtencion", {
-                required: "Centro requerido",
-              })}
-              className="form-input"
-            >
-              <option value="">-</option>
-              {centros.map((ca, index) => (
-                <option key={index} value={ca.id}>
-                  {ca.nombre + " - " + ca.localidad?.denominacion + ", " + ca.domicilio}
-                </option>
-              ))}
-            </select>
-            {errorsAdd.centroAtencion && (
-              <div className="error-message">{errorsAdd.centroAtencion.message}</div>
-            )}
-          </div>
-
-          <div className="form-group" style={{gridColumn: "1", gridRow : "3"}}>
-            <label htmlFor="file">Receta</label>
-            <input type="file"
-              accept="image/*"
-              name="receta"
-              {...registerAdd("receta",
-                {
-                  required: "Receta requerida",
-                  validate: {
-                    isImage: (value) => {
-                      if (!value[0]) {
-                        return true;
-                      }
-                      const fileType = value[0].type;
-                      const acceptedImageTypes = ['image/jpeg', 'image/jpg', 'image/png', 'image/bmp'];
-                      return acceptedImageTypes.includes(fileType) || "El archivo debe ser una imagen (JPG, PNG).";
-                    },
-                    isSize: (value) => {
-                      if (value[0].size < 10 * 1024 * 1024) {
-                        return true;
-                      }
-                      return "El archivo debe ser menor a 10mb.";
-                    }
-                  }
-                })}
-            />
-            {errorsAdd.receta && (
-              <div className="error-message">{errorsAdd.receta.message}</div>
-            )}
-          </div>
-          <div className="form-options" style={{gridColumn: "2", gridRow : "3"}}>
-            <label className="checkbox-label">
-              <input type="checkbox" {...registerAdd("recibeMail")} />
-              <span>Deseo recibir Email recordatorio</span>
-            </label>
-          </div>
-          <button id="turno" type="submit" className="login-btn" disabled={isSubmittingAdd} style={{gridColumn: "2", gridRow : "4"}}>
-            {isSubmittingAdd ? "Un momento..." : "Registrar"}
-          </button>
-        </form>
+        <TabTurnoRegistrar tipos={tipos} centros={centros} registerAdd={registerAdd} handleSubmitAdd={handleSubmitAdd} onSubmitAdd={onSubmitAdd} errorsAdd={errorsAdd} isSubmittingAdd={isSubmittingAdd} horariosDisponibles={horariosDisponibles} watchFecha={fechaHoraReserva} />
       </Tab>
       <Tab eventKey="resultados" title="Resultados">
-        <h2 className='titulo'>Resultados disponibles</h2>
-        {turnosFiltradosResultados.length > 0 ? (
-          <div style={{ marginTop: '20px' , overflowX: 'scroll'}}>
-            <table className="table" style={{ verticalAlign: 'middle' }}>
-              <thead>
-                <tr>
-                  <th>Número de Turno</th>
-                  <th>Tipo de Análisis</th>
-                  <th>Centro de Atención</th>
-                  <th>Fecha y Hora</th>
-                  <th>Estado</th>
-                  <th>Observación</th>
-                  <th>Recibe Mail</th>
-                  <th>Mostrar Resultados</th>
-                </tr>
-              </thead>
-              <tbody>
-                {turnosFiltradosResultados.map((turno) => {
-                  if (turno.estado !== "Resultado") {
-                    return null;
-                  }
-                  return (
-                    <tr key={turno.id}>
-                      <td>{turno.id}</td>
-                      <td>{turno.tipoAnalisis?.nombre || '-'}</td>
-                      <td>{turno.centroAtencion?.nombre || '-'}</td>
-                      <td>{new Date(turno.fechaHoraReserva).toLocaleString()}</td>
-                      <td>{turno.estado}</td>
-                      <td>{turno.observacion === "" ? "-" : turno.observacion}</td>
-                      <td>{turno.recibeMail ? "Sí" : "No"}</td>
-                      <td>
-                        <button
-                        onClick={() => handleResultadosClick(turno.id)}
-                          style={{
-                            background: 'none',
-                            border: 'none',
-                            color: 'green',
-                            cursor: 'pointer',
-                            fontSize: '20px',
-                            fontStyle: 'underline'
-                          }}
-                        >
-                          ✔️
-                        </button>
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          </div>
-        ) : (
-          <p style={{ marginTop: "30px" }}>No tienes resultados disponibles.</p>
-        )}
-        {showModal && (
-          <div style={{
-            position: 'fixed',
-            top: 0,
-            left: 0,
-            width: '100%',
-            height: '100%',
-            backgroundColor: 'rgba(0, 0, 0, 0.5)',
-            display: 'flex',
-            justifyContent: 'center',
-            alignItems: 'center',
-            zIndex: 1000,
-          }}>
-            <div style={{
-              backgroundColor: 'white',
-              padding: '20px',
-              borderRadius: '5px',
-              textAlign: 'center',
-              minWidth: '300px'
-            }}>
-              <h4 style={{fontWeight: 'bold', fontFamily: 'Segoe UI, Tahoma, Geneva, Verdana, sans-serif'}}>
-                Resultados</h4>
-                {resultadosId && (
-                  <div style={{display: 'grid', gridTemplateColumns: '1fr 1fr', columnGap: "50px", rowGap: "20px", marginTop: "20px"}}>
-                    {turnosPaciente.find((t) => t.id === Number(resultadosId))
-                      .resultados.map((resultado) => (
-                        <div key={resultado.parametroAnalisis.id} className="form-group">
-                          <label style={{fontWeight:"bold", fontSize:"1.2rem"}}>
-                            {resultado.parametroAnalisis.nombre} ({resultado.parametroAnalisis.unidad})
-                          </label>
-                          <label style={{fontWeight:"bold", fontSize:"1.2rem"}}>{resultado.valor}</label>
-                           <label>Referencia: {resultado.parametroAnalisis.referencia} {resultado.parametroAnalisis.unidad}</label>
-                        </div>
-                      ))}
-                  </div>
-              )}
-              <div style={{ marginTop: '20px' }}>
-                <button onClick={handleCerrarModal} className='login-btn' style={{ backgroundColor: 'red' }}>
-                  Volver
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
+        <TabTurnoResultados turnosFiltradosResultados={turnosFiltradosResultados} onResultadosClick={handleResultadosClick} />
       </Tab>
     </Tabs>
   );
