+++ frontend/src/components/tabsUsers/TabLaboratorioPresupuesto.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const TabLaboratorioPresupuesto = () => {
+  return (
+    <>
+      <h2 className='titulo'>Presupuestos</h2>
+      <p>A implementar...</p>
+    </>
+  );
+};
+
+export default TabLaboratorioPresupuesto;
