+++ frontend/src/hooks/useCentros.js
@@ -27,7 +27,6 @@ const addCentros = async (data) => {
   };
   try {
     const response = await axiosInstance.post('/centroAtencion', centrosData);
-    console.log(response);
     alert('Centro n°:' + response.data.data.id + ' creado correctamente');
   } catch (error) {
     console.error('Error al crear los datos:', error);
